[Association] New behat scenarios
