+++ behat.yml.dist
@@ -219,8 +219,8 @@ default:
                 - Sylius\Bundle\CoreBundle\Behat\CoreContext
                 - Sylius\Bundle\MoneyBundle\Behat\MoneyContext
                 - Sylius\Bundle\ProductBundle\Behat\ProductContext
+                - Sylius\Bundle\ProductBundle\Behat\ProductAssociationContext
                 - Sylius\Bundle\ResourceBundle\Behat\FixtureContext
-                - Sylius\Bundle\AssociationBundle\Behat\ProductAssociationContext
                 - Sylius\Bundle\TaxationBundle\Behat\TaxationContext
                 - Sylius\Bundle\TaxonomyBundle\Behat\TaxonomyContext
                 - Sylius\Bundle\WebBundle\Behat\WebContext
